extract web view into separated widget
clean HomeScreen.dart file
